feat: add has_manifest_ref and renamed project_ref
