+++ crates/pixi_manifest/src/has_manifest_ref.rs
@@ -1,0 +1,4 @@
+pub trait HasManifestRef<'m> {
+    /// Returns access to the original pixi manifest
+    fn manifest(&self) -> &'m crate::Manifest;
+}
