+++ src/project/grouped_environment.rs
@@ -8,7 +8,7 @@ use crate::{
 };
 use itertools::Either;
 use pixi_consts::consts;
-use pixi_manifest::{EnvironmentName, Feature, Manifest, SystemRequirements};
+use pixi_manifest::{EnvironmentName, Feature, HasManifestRef, Manifest, SystemRequirements};
 use rattler_conda_types::{GenericVirtualPackage, Platform};
 use std::path::PathBuf;
 
@@ -119,11 +119,13 @@ impl<'p> HasProjectRef<'p> for GroupedEnvironment<'p> {
     }
 }
 
-impl<'p> HasFeatures<'p> for GroupedEnvironment<'p> {
+impl<'p> HasManifestRef<'p> for GroupedEnvironment<'p> {
     fn manifest(&self) -> &'p Manifest {
         &self.project().manifest
     }
+}
 
+impl<'p> HasFeatures<'p> for GroupedEnvironment<'p> {
     /// Returns the features of the group
     fn features(&self) -> impl DoubleEndedIterator<Item = &'p Feature> + 'p {
         match self {
