+++ src/project/mod.rs
@@ -2,7 +2,7 @@ mod environment;
 pub mod errors;
 pub mod grouped_environment;
 pub mod has_features;
-mod project_ref;
+mod has_project_ref;
 mod repodata;
 mod solve_group;
 pub mod virtual_packages;
@@ -42,7 +42,7 @@ use crate::{
 use pixi_config::Config;
 use pixi_consts::consts;
 
-pub use project_ref::HasProjectRef;
+pub use has_project_ref::HasProjectRef;
 
 static CUSTOM_TARGET_DIR_WARN: OnceCell<()> = OnceCell::new();
